Omitted unnecessary checks if admin type remain the same

+++ editadminuser.php
@@ -21,7 +21,7 @@ if ($ref!=$_SERVER['SCRIPT_FILENAME']){
 	<?php
 	if ($admintype == 0){
 		$handle1=DB::connect($DB['DSN'], true);
-		$query = "SELECT * from adminuser WHERE username='$username'";
+		$query = "SELECT `type` from adminuser WHERE username='$username'";
 		$result = $handle1->query($query);
 		$cnt = $result->numRows($result);
 		$adminrow = $result->fetchRow(DB_FETCHMODE_ASSOC, 0);
@@ -195,10 +195,11 @@ if ($ref!=$_SERVER['SCRIPT_FILENAME']){
 			}
 			elseif (! empty($confirmed)){
 
-				# First delete and set new Domainresponsibilities
-
-				$query = "DELETE from domainadmin WHERE adminuser='$username'";
-				$result = $handle1->query($query);
+				if ($type != $newtype || $type = '1') {
+					# First delete and set new Domainresponsibilities
+					$query = "DELETE from domainadmin WHERE adminuser='$username'";
+					$result = $handle1->query($query);
+				}
 
 				# Insert each key in the array "domain" into database again
 				if (isset($resp_domain) && sizeof($Resp_domain)>0) {
@@ -210,7 +211,7 @@ if ($ref!=$_SERVER['SCRIPT_FILENAME']){
 
 				# If there is a new domain to add, lets insert it to the DB
 
-				if (isset($newdomain)){
+				if ($newdomain){
 
 					# Check if the domain to be added really exists
 
@@ -254,44 +255,46 @@ if ($ref!=$_SERVER['SCRIPT_FILENAME']){
 				}
 
 				# The only admin user related thing we can do now is updating the type of the admin
+				if ($type != $newtype) {
+					if ($newtype == 1){
 
-				if ($newtype == 1){
+						# We have to take care to have at least one superuser left, or we cannot use
+						# Web-cyradm again
 
-					# We have to take care to have at least one superuser left, or we cannot use
-					# Web-cyradm again
+						# Query to get the count of superusers
+						$query="SELECT type FROM adminuser WHERE type='0'";
+						$result = $handle1->query($query);
+						$cnt=$result->numRows($result);
+						
+						# This is not necessary! $type is already set
+						# Determine what kind of user is beeing edited
+						$query="SELECT type FROM adminuser WHERE username='$username'";
+						$result=$handle1->query($query);
+						$row = $result->fetchRow(DB_FETCHMODE_ASSOC, 0);
+						$type= $row['type'];
 
-					# Query to get the count of superusers
-					$query="SELECT type FROM adminuser WHERE type='0'";
-					$result = $handle1->query($query);
-					$cnt=$result->numRows($result);
-
-					# Determine what kind of user is beeing edited
-					$query="SELECT type FROM adminuser WHERE username='$username'";
-					$result=$handle1->query($query);
-					$row = $result->fetchRow(DB_FETCHMODE_ASSOC, 0);
-					$type= $row['type'];
-
-					# Check if only 1 superuser is defined, in case of requested change of a superuser
-					if ($cnt==1 && $type==0){
-						# No Way! We cannot change the last Superuser to domainadmin!
-						die (_("At least one Superuser is needed for Web-cyradm"));
+						# Check if only 1 superuser is defined, in case of requested change of a superuser
+						if ($cnt==1 && $type==0){
+							# No Way! We cannot change the last Superuser to domainadmin!
+							die (_("At least one Superuser is needed for Web-cyradm"));
+						}
+						# If not died, lets update the type
+						$query="UPDATE adminuser SET type='$newtype' WHERE username='$username'";
+						$result=$handle1->query($query);
 					}
-					# If not died, lets update the type
-					$query="UPDATE adminuser SET type='$newtype' WHERE username='$username'";
-					$result=$handle1->query($query);
-				}
-				if ($newtype==0){
+					if ($newtype==0){
 
-					# If the new type of admin will be superuser, we also need to delete the domain
-					# Responsibilities, because a superuser is always responsible for all domains
-					$query = "DELETE FROM domainadmin WHERE adminuser='$username'";
-					$result = $handle1->query($query);
+						# If the new type of admin will be superuser, we also need to delete the domain
+						# Responsibilities, because a superuser is always responsible for all domains
+						$query = "DELETE FROM domainadmin WHERE adminuser='$username'";
+						$result = $handle1->query($query);
 
-					# And update the type of this admin
-					$query="UPDATE adminuser SET type='$newtype' WHERE username='$username'";
-					$result=$handle1->query($query);
-					$query="INSERT INTO domainadmin (domain_name,adminuser) VALUES ('*','$username')";
-  				 	$result=$handle1->query($query);
+						# And update the type of this admin
+						$query="UPDATE adminuser SET type='$newtype' WHERE username='$username'";
+						$result=$handle1->query($query);
+						$query="INSERT INTO domainadmin (domain_name,adminuser) VALUES ('*','$username')";
+  					 	$result=$handle1->query($query);
+					}
 				}
 
 				include WC_BASE . "/adminuser.php";
